Extracted parameter updates from QueryParameters
